LayoutTests/platform/chromium/inspector/debugger/watch-expressions-panel-switch-expected.txt

Tests debugger does not fail when stopped while a panel other than scripts was opened. Both valid and invalid expressions are added to watch expressions.

Bug 70718
Debugger was enabled.
Script source was shown.
Set timer for test function.
Watch expressions updated.
[expanded]  
    window.document: #document
    windowa.document: ReferenceError: windowa is not defined
Debugger was disabled.
